+++ Sphinx/ChangeLog.rst
@@ -2,6 +2,7 @@ ChangeLog
 =========
 * May 23, 2014
 
+  - Off the mixer if no sound is loaded.
   - Updated ``example/RiceRocks_Asteroids.py``.
 
   - Updated ``example/Spaceship_prototype.py``.